Fix broken link to EN documentation
The link pointing to the EN documentation was showing a 403 page. I found the correct one through the Japanese version under Translations, only a ""#" was missing.

Thank you for this awesome project and i am looking forward to playing around with it and hopefully maybe helping you out if i can.